Add trackHistory and TrackHistoryWithMatomo
